requirements.txt
boto3>=1.34.0
pandas>=2.2.0
numpy>=1.26.0
python-dotenv>=1.0.1
